Added required fields of the TraineeController.

+++ src/main/java/com/javacore/task/controllers/TraineeController.java
@@ -1,12 +1,12 @@
 package com.javacore.task.controllers;
 
 import com.javacore.task.models.TraineeModel;
-import com.javacore.task.models.response.TraineeProfileUpdateResponse;
 import com.javacore.task.models.request.TraineeTrainingsRequest;
 import com.javacore.task.models.request.TraineeUpdateRequest;
 import com.javacore.task.models.response.TraineeInfoResponse;
-import com.javacore.task.models.response.TrainersListResponse;
+import com.javacore.task.models.response.TraineeProfileUpdateResponse;
 import com.javacore.task.models.response.TraineeTrainingInfoResponse;
+import com.javacore.task.models.response.TrainersListResponse;
 import com.javacore.task.services.TraineeService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
@@ -34,7 +34,7 @@ public class TraineeController {
 
     @Operation(description = "Get Trainee by ID")
     @ApiResponses(value = {
-            @ApiResponse(responseCode = "200", description= "Successfully retrieved trainee"),
+            @ApiResponse(responseCode = "200", description = "Successfully retrieved trainee"),
             @ApiResponse(responseCode = "404", description = "Trainee not found")
     })
     @GetMapping("/{traineeId}")
@@ -46,9 +46,9 @@ public class TraineeController {
         return ResponseEntity.ok(traineeModel);
     }
 
-    @Operation(description = "Update Trainee")
+    @Operation(description = "Update Trainee profile")
     @ApiResponses(value = {
-            @ApiResponse(responseCode = "200", description= "Successfully updated trainee"),
+            @ApiResponse(responseCode = "200", description = "Successfully updated trainee profile"),
             @ApiResponse(responseCode = "404", description = "Trainee not found")
     })
     @PutMapping
@@ -59,15 +59,17 @@ public class TraineeController {
         log.info("Response: {}", updatedTrainee);
         return ResponseEntity.ok(updatedTrainee);
     }
-
     @Operation(description = "Delete Trainee")
     @ApiResponses(value = {
-            @ApiResponse(responseCode = "200", description= "Successfully deleted trainee"),
+            @ApiResponse(responseCode = "200", description = "Successfully deleted trainee"),
             @ApiResponse(responseCode = "404", description = "Trainee not found")
     })
     @DeleteMapping
     public ResponseEntity<String> deleteTrainee(@RequestParam String username) {
-        log.info("Endpoint called: DELETE /trainees?username={}", username );
+        if (username == null || username.trim().isEmpty()) {
+            throw new IllegalArgumentException("Trainee username cannot be null or empty");
+        }
+        log.info("Endpoint called: DELETE /trainees?username={}", username);
         traineeService.deleteTrainee(username);
         log.info("Response: deleted successfully");
         return new ResponseEntity<>("deleted successfully", HttpStatus.OK);
@@ -75,7 +77,7 @@ public class TraineeController {
 
     @Operation(description = "Get Trainee Profile")
     @ApiResponses(value = {
-            @ApiResponse(responseCode = "200", description= "Successfully retrieved trainee profile"),
+            @ApiResponse(responseCode = "200", description = "Successfully retrieved trainee profile"),
             @ApiResponse(responseCode = "404", description = "Trainee not found")
     })
     @GetMapping
@@ -92,12 +94,15 @@ public class TraineeController {
 
     @Operation(description = "Update Trainee Status")
     @ApiResponses(value = {
-            @ApiResponse(responseCode = "200", description= "Successfully updated trainee status"),
+            @ApiResponse(responseCode = "200", description = "Successfully updated trainee status"),
             @ApiResponse(responseCode = "404", description = "Trainee not found"),
             @ApiResponse(responseCode = "409", description = "Trainee is already in desired state")
     })
     @PatchMapping
-    public ResponseEntity<String> updateTraineeStatus(@RequestParam String username, @RequestParam("status") boolean status) {
+    public ResponseEntity<String> updateTraineeStatus(@RequestParam String username, @RequestParam("status") Boolean status) {
+        if ((username == null || username.trim().isEmpty()) || (status == null)) {
+            throw new IllegalArgumentException("Trainee username or status cannot be null or empty");
+        }
         log.info("Endpoint called: PATCH /trainees?username={}&choice={}", username, status);
         String result = traineeService.updateTraineeStatus(status, username);
         log.info("Response: {}", result);
@@ -107,12 +112,15 @@ public class TraineeController {
 
     @Operation(description = "Get Not Assigned Trainers")
     @ApiResponses(value = {
-            @ApiResponse(responseCode = "200", description= "Successfully retrieved not assigned trainers"),
+            @ApiResponse(responseCode = "200", description = "Successfully retrieved not assigned trainers"),
             @ApiResponse(responseCode = "404", description = "Trainers not found")
     })
     @GetMapping("/not-assigned-trainers")
     @PreAuthorize("hasAuthority('TRAINEE')")
     public ResponseEntity<List<TrainersListResponse>> getNotAssignedTrainersForTrainee(@RequestParam String username) {
+        if (username == null || username.trim().isEmpty()) {
+            throw new IllegalArgumentException("Trainee username cannot be null or empty");
+        }
         log.info("Endpoint called: GET /not-assigned-trainers?username={}", username);
         List<TrainersListResponse> responses = traineeService.getNotAssignedActiveTrainersListForTrainee(username);
         log.info("Response: {}", responses);
@@ -121,11 +129,14 @@ public class TraineeController {
 
     @Operation(description = "Update Trainee Trainers List")
     @ApiResponses(value = {
-            @ApiResponse(responseCode = "200", description= "Successfully updated trainee trainers list"),
+            @ApiResponse(responseCode = "200", description = "Successfully updated trainee trainers list"),
             @ApiResponse(responseCode = "404", description = "Trainers not found")
     })
     @PutMapping("/update-trainers")
     public ResponseEntity<List<TrainersListResponse>> updateTraineeTrainersList(@RequestParam("username") String username, @RequestBody List<String> trainersUsernames) {
+        if ((username == null || username.trim().isEmpty()) || trainersUsernames.isEmpty()) {
+            throw new IllegalArgumentException("Trainee username or trainers usernames list cannot be null or empty");
+        }
         log.info("Endpoint called: PUT /update-trainers?username={}, REQUEST: trainersUsernames={}", username, trainersUsernames);
         List<TrainersListResponse> responses = traineeService.updateTraineeTrainersList(username, trainersUsernames);
         log.info("Response: {}", responses);
@@ -134,7 +145,7 @@ public class TraineeController {
 
     @Operation(description = "Get Trainee Trainings List")
     @ApiResponses(value = {
-            @ApiResponse(responseCode = "200", description= "Successfully retrieved trainee trainings list"),
+            @ApiResponse(responseCode = "200", description = "Successfully retrieved trainee trainings list"),
             @ApiResponse(responseCode = "404", description = "Trainings not found")
     })
     @PostMapping("/trainee-trainings")
